refactor(radio): isolate request methods to channels
